// api/src/core/repositories/ProductRepository/dto/addProductDto.d.ts
export declare class AddProductDto {
    readonly id: string;
    readonly name: string;
    readonly manufacturer: string;
    readonly category: string;
    readonly subCategory: string;
    readonly compatibility: string;
    readonly description: string;
    readonly availability: boolean;
    readonly price: number;
    constructor(id: string, name: string, manufacturer: string, category: string, subCategory: string, compatibility: string, description: string, availability: boolean, price: number);
}
//# sourceMappingURL=addProductDto.d.ts.map